add licences of vendored code
